0.8.9 R2 & 0.9 Pre-Release 4 Update
-----R2 (2A\1C)-----
-Fixed A MILLION BUGS RELATED TO THE DAMN HW IDENTIFIER.
-Added alternative to 'exit' - 'shutdown'.
-Fixed BIOS (I REPEAT, DO NOT RUN THE CUSTOM AWARD BIOS!)
-Made it so hard disks aren't mandatory anymore. (floppy drives take its place).
-Added OEM Computers (Macintosh, IBM, etc. Basically, a single file containing all the information of HW, plus the PC's model name.).
-Added a BIOS lock ;).
(Due to the size of the update and countless identifier fixes, this update will be released both as '0.8.9 R2' and '0.9 Pre-Release 4', because I don't know for how long or since
what version those bugs existed. This update was supposed to be release of 0.9, but because of those bugs, I'll delay it even further. Don't worry, there will be some tiny extra
features coming in 0.9, as well as updated READMEs and Guides, so lock in!)